+++ test/fixtures/class-generics/actual.js
@@ -1,5 +1,11 @@
 class A<T> {
-  constructor(x: T) {
-    this.x = x;
-  }
+  foo(x: T) {}
+}
+
+class B extends C<T> {
+  foo(x: T) {}
+}
+
+class D<T: F> extends E<T> {
+  foo(x: T) {}
 }