+++ test/fixtures/interface/expected.js
@@ -15,3 +15,8 @@ const D = _extend([A, _t.refinement(_t.interface({}), p)], "D");
 const E = _extend([A, {
   e: _t.Any
 }], "E");
+
+const F = _extend([E, {
+  e: _t.Any,
+  f: _t.Any
+}], "F");